Moving from old als to new class-based als and arpls
